requirements.txt
requests
python-dotenv
aiohttp
requests
urllib3
uvicorn
fastapi
pydantic
httpx
sqlalchemy
databases
psycopg2-binary
asyncpg
python-multipart
email-validator
python-telegram-bot
jinja2




